Update job execution to Micah's suggestion
