+++ source/SIL.AppBuilder.Portal/node-server/job-executors/base.ts
@@ -1,0 +1,7 @@
+import { BullMQ } from 'sil.appbuilder.portal.common';
+import { Job } from 'bullmq';
+
+export abstract class ScriptoriaJobExecutor<T extends BullMQ.ScriptoriaJobType> {
+  constructor() {}
+  abstract execute(job: Job<BullMQ.JobTypeMap[T], number, string>): Promise<number>;
+}
