+++ source/SIL.AppBuilder.Portal/node-server/job-executors/index.ts
@@ -1,0 +1,2 @@
+export { Test } from './test.js';
+export { ReassignUserTasks } from './reassignUserTasks.js';
